+++ composer.json
@@ -1,10 +1,10 @@
 {
-    "name": "bjyoungblood/bjy-profiler",
-    "description": "Database profiler for Zend\\Db (also plugin for ZendDeveloperTools)",
+    "name": "andrey-mokhov/bjy-profiler",
+    "description": "Database profiler for Zend\\Db (also plugin for ZendDeveloperTools). Module forked from bjyoungblood/bjy-profiler",
     "type": "library",
     "license": "MIT",
     "keywords": [
-        "zf2",
+        "zf3",
         "db",
         "profiler",
         "zdt"
@@ -15,18 +15,26 @@
             "name": "Ben Youngblood",
             "email": "bx.youngblood@gmail.com",
             "homepage": "http://bjyoungblood.com/"
+        },
+        {
+            "name": "Andrey N. Mokhov",
+            "email": "github@mokhov.com",
+            "homepage": "https://github.com/andrey-mokhov"
         }
     ],
     "require": {
-        "php": ">=5.3",
-        "zendframework/zend-db": ">=2.1.0"
+        "php": ">=5.5",
+        "zendframework/zend-db": ">=2.8.0"
     },
     "autoload": {
-        "psr-0": {
-            "BjyProfiler": "src/"
-        },
-        "classmap": [
-            "./Module.php"
-        ]
+        "psr-4": {
+            "BjyProfiler\\": "src/"
+        }
+    },
+    "extra": {
+        "zf": {
+            "component": "BjyProfiler",
+            "config-provider": "BjyProfiler\\ConfigProvider"
+        }
     }
 }
