--- src/BjyProfiler/Db/Adapter/ProfilingAdapterFactory.php
@@ -1,36 +1,0 @@
-<?php
-/**
- * Created by Inditel Meedia OÜ
- * User: Oliver Leisalu
- */
-
-namespace BjyProfiler\Db\Adapter;
-
-
-use BjyProfiler\Db\Profiler\Profiler;
-use Zend\ServiceManager\FactoryInterface;
-use Zend\ServiceManager\ServiceLocatorInterface;
-
-class ProfilingAdapterFactory implements FactoryInterface
-{
-
-    /**
-     * @param ServiceLocatorInterface $serviceLocator
-     * @return mixed
-     */
-    public function createService(ServiceLocatorInterface $serviceLocator)
-    {
-        $config = $serviceLocator->get('Configuration');
-        $dbParams = $config['db'];
-        $adapter = new ProfilingAdapter($dbParams);
-
-        $adapter->setProfiler(new Profiler);
-        if (isset($dbParams['options']) && is_array($dbParams['options'])) {
-            $options = $dbParams['options'];
-        } else {
-            $options = array();
-        }
-        $adapter->injectProfilingStatementPrototype($options);
-        return $adapter;
-    }
-}
