+++ src/Module.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace BjyProfiler;
+
+class Module
+{
+    /**
+     * Retrieve default zend-db configuration for zend-mvc context.
+     *
+     * @return array
+     */
+    public function getConfig()
+    {
+        $provider = new ConfigProvider();
+        return [
+            'service_manager' => $provider->getDependencyConfig(),
+        ];
+    }
+}
